move default props to top of file, include smallText and add another test
